requirements.txt
langchain
langchain-openai
langchain-community
langchain-huggingface
sentence-transformers
huggingface_hub
faiss-cpu
openai
pypdf
streamlit
rich
tf-keras
pillow
pytesseract